Add option to not pair MSAs in run_multimer_jobs.py
The flag `--no_pair_msa`, allows the user to prevent the multimer MSA from being paired between chains. This is particularly useful for cases where the user intends to model the interaction between a protein and fragments of itself. This represents a homomeric interaction, but will not be recognized as such by `_is_homomer_or_monomer` as the fragment has a different length than the full protein.

Made `pair_and_merge` a method of `MultimericObject`, so that pairing could be controlled by the class variable `pair_msa`.

Removed unecessary calls to `create_all_chain_features`, which is already called by the constructor